logic done

+++ app/src/main/java/com/example/weather/data/dto/Clouds.kt
@@ -1,9 +1,10 @@
 package com.example.weather.data.dto
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 
 data class Clouds (
-
-	@SerializedName("all") val all : Int
+	@SerializedName("all")
+	val all : BigDecimal
 )
