+++ app/src/main/java/com/example/weather/data/dto/Coord.kt
@@ -1,9 +1,12 @@
 package com.example.weather.data.dto
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class Coord (
 
-	@SerializedName("lon") val lon : Double,
-	@SerializedName("lat") val lat : Double
+	@SerializedName("lon")
+	val lon : BigDecimal,
+	@SerializedName("lat")
+	val lat : BigDecimal
 )
