+++ app/src/main/java/com/example/weather/data/dto/Weather.kt
@@ -1,11 +1,15 @@
 package com.example.weather.data.dto
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class Weather (
-
-	@SerializedName("id") val id : Int,
-	@SerializedName("main") val main : String,
-	@SerializedName("description") val description : String,
-	@SerializedName("icon") val icon : String
+	@SerializedName("id")
+	val id : BigDecimal,
+	@SerializedName("main")
+	val main : String,
+	@SerializedName("description")
+	val description : String,
+	@SerializedName("icon")
+	val icon : String
 )
